Move BackoffPolicy to common package
