+++ server/src/main/java/org/elasticsearch/common/BackoffPolicy.java
@@ -6,7 +6,7 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
-package org.elasticsearch.action.bulk;
+package org.elasticsearch.common;
 
 import org.elasticsearch.core.TimeValue;
